moved extension test to extension folder
